Fix checkstyle.
